Moving the Modal molecule to the directory: src/components/molecules, Modal.test: to the directory /src/__tes__/components/molecules,
